[FIX] 修复手机号注册 登陆JWT TOKEN出现问题

+++ generator-map/src/main/java/com/nexus/manager/pojo/TbItem.java
@@ -6,7 +6,7 @@ import java.util.Date;
 public class TbItem {
     private Long id;
 
-    private Integer categoryId;
+    private Long categoryId;
 
     private String title;
 
@@ -26,7 +26,7 @@ public class TbItem {
 
     private Date updateTime;
 
-    public TbItem(Long id, Integer categoryId, String title, String subtitle, String mainImage, String subImage, BigDecimal price, Integer stock, String detail, Date createTime, Date updateTime) {
+    public TbItem(Long id, Long categoryId, String title, String subtitle, String mainImage, String subImage, BigDecimal price, Integer stock, String detail, Date createTime, Date updateTime) {
         this.id = id;
         this.categoryId = categoryId;
         this.title = title;
@@ -52,11 +52,11 @@ public class TbItem {
         this.id = id;
     }
 
-    public Integer getCategoryId() {
+    public Long getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(Integer categoryId) {
+    public void setCategoryId(Long categoryId) {
         this.categoryId = categoryId;
     }
